consumer/requirements.txt
apscheduler
kafka-python
requests
